timespan HH++

+++ Core/xStatic/xExtension/ExtTimeSpan.cs
@@ -9,7 +9,7 @@ namespace xLib
 		{
 			string temp = format;
 			temp = temp.Replace("dd",timeSpan.Days.ToString("00"));
-			temp = temp.Replace("HH",timeSpan.TotalHours.ToString("00"));
+			temp = temp.Replace("HH",((int)timeSpan.TotalHours).ToString("00"));
 			temp = temp.Replace("hh",timeSpan.Hours.ToString("00"));
 			temp = temp.Replace("mm",timeSpan.Minutes.ToString("00"));
 			temp = temp.Replace("ss",timeSpan.Seconds.ToString("00"));
